more 64-bit madness
git-svn-id: http://svn.macosforge.org/repository/ruby/MacRuby/branches/macruby64@671 23306eb0-4c56-4727-a40e-e92c0eb68959

+++ thread.c
@@ -510,7 +510,7 @@ thread_initialize(VALUE thread, VALUE args)
 		     file);
 	}
         rb_raise(rb_eThreadError, "already initialized thread - %s:%d",
-                 file, NUM2INT(line));
+                 file, (int)NUM2INT(line));
     }
     return thread_create_core(thread, args, 0);
 }
